+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## X.Y.Z 2023-y-z
 ### PaymentSheet
 * [Fixed] Visual bug where re-presenting PaymentSheet wouldn't show a spinner while it reloads.
+* [Added] If PaymentSheet fails to load a deferred intent configuration, we fall back to displaying cards (or the intent configuration payment method types) instead of failing.
 
 ## 23.18.2 2023-11-06
 ### CustomerSheet